Reverted the type used to represent codimensions from size_t to int.
Arguments:
- codimensions are not sizes
- in Dune there asserts like "assert(codim >= 0)" designed to spot errors
  resulting from substraction of constants from codimensions. These
  would be defeated if codimensions were represented with size_t
  (since the 0 - 1 == 2^64 - 1 > 0).

+++ lib/assembly/grid_function.hpp
@@ -52,7 +52,7 @@ namespace Bempp
 class AssemblyOptions;
 class GeometryFactory;
 class Grid;
-template <size_t codim> class Entity;
+template <int codim> class Entity;
 template <typename BasisFunctionType> class Space;
 
 template <typename BasisFunctionType, typename ResultType>
