+++ lib/grid/concrete_range_entity_iterator.hpp
@@ -74,7 +74,7 @@ public:
     }
 
     virtual std::auto_ptr<EntityPointer<ConcreteRangeEntityIterator::codimension> > frozen() const {
-        const size_t codim = ConcreteRangeEntityIterator::codimension;
+        const int codim = ConcreteRangeEntityIterator::codimension;
         return std::auto_ptr<EntityPointer<codim> >(
                     new ConcreteEntityPointer<DuneEntityPointer>(*m_cur));
     }
